feat: removed MerklClaimHandler
